requirements.txt
actionlib==1.14.3
angles==1.9.14
Backtesting==0.4.0
bokeh==3.1.1
bondpy==1.8.7
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
certifi==2025.10.5
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
dynamic-reconfigure==1.7.6
exchange-calendars==4.2.8
gazebo-plugins==2.9.3
gazebo-ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
image-geometry==1.16.2
interactive-markers==1.12.2
jinja2==3.1.6
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
korean-lunar-calendar==0.3.1
laser-geometry==1.6.8
MarkupSafe==2.1.5
message-filters==1.17.4
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pandas-market-calendars==5.1.1
pillow==10.4.0
polygon-api-client==1.15.4
pyluach==2.3.0
python-dateutil==2.9.0.post0
python-qt-binding==0.4.6
pytz==2025.2
PyYAML==6.0.3
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
resource-retriever==1.12.10
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rviz==1.14.26
sensor-msgs==1.13.2
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
toolz==1.0.0
topic-tools==1.17.4
tornado==6.4.2
tzdata==2025.2
urllib3==2.2.3
websockets==13.1
xacro==1.14.20
xyzservices==2025.4.0
